+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/base/ImageTag.java
@@ -234,7 +234,7 @@ public abstract class ImageTag extends DrawableTag {
 
     @Override
     public void toImage(int frame, int time, int ratio, RenderContext renderContext, SerializableImage image, SerializableImage fullImage, boolean isClip, Matrix transformation, Matrix strokeTransformation, Matrix absoluteTransformation, Matrix fullTransformation, ColorTransform colorTransform, double unzoom, boolean sameImage, ExportRectangle viewRect, boolean scaleStrokes, int drawMode) {
-        BitmapExporter.export(1, swf, getShape(), null, image, transformation, strokeTransformation, colorTransform, true);
+        BitmapExporter.export(1, swf, getShape(), null, image, unzoom, transformation, strokeTransformation, colorTransform, true);
     }
 
     @Override